+++ client/include/webcface/image.h
@@ -49,8 +49,12 @@ class WEBCFACE_DLL Image : protected Field {
     }
     /*!
      * \since ver1.11
+     * \deprecated ver2.8〜
      */
-    Image child(int index) const { return this->Field::child(index); }
+    [[deprecated]]
+    Image child(int index) const {
+        return this->Field::child(std::to_string(index));
+    }
     /*!
      * child()と同じ
      * \since ver1.11
@@ -73,8 +77,12 @@ class WEBCFACE_DLL Image : protected Field {
     /*!
      * child()と同じ
      * \since ver1.11
+     * \deprecated ver2.8〜
      */
-    Image operator[](int index) const { return child(index); }
+    [[deprecated]]
+    Image operator[](int index) const {
+        return child(std::to_string(index));
+    }
     /*!
      * \brief nameの最後のピリオドの前までを新しい名前とするField
      * \since ver1.11