+++ client/include/webcface/robot_model.h
@@ -56,8 +56,12 @@ class WEBCFACE_DLL RobotModel : protected Field {
     }
     /*!
      * \since ver1.11
+     * \deprecated ver2.8〜
      */
-    RobotModel child(int index) const { return this->Field::child(index); }
+    [[deprecated]]
+    RobotModel child(int index) const {
+        return this->Field::child(std::to_string(index));
+    }
     /*!
      * child()と同じ
      * \since ver1.11
@@ -82,8 +86,12 @@ class WEBCFACE_DLL RobotModel : protected Field {
     /*!
      * child()と同じ
      * \since ver1.11
+     * \deprecated ver2.8〜
      */
-    RobotModel operator[](int index) const { return child(index); }
+    [[deprecated]]
+    RobotModel operator[](int index) const {
+        return child(std::to_string(index));
+    }
     /*!
      * \brief nameの最後のピリオドの前までを新しい名前とするField
      * \since ver1.11