+++ client/include/webcface/view.h
@@ -70,8 +70,12 @@ class WEBCFACE_DLL View : protected Field {
     }
     /*!
      * \since ver1.11
+     * \deprecated ver2.8〜
      */
-    View child(int index) const { return this->Field::child(index); }
+    [[deprecated]]
+    View child(int index) const {
+        return this->Field::child(std::to_string(index));
+    }
     /*!
      * child()と同じ
      * \since ver1.11
@@ -94,8 +98,12 @@ class WEBCFACE_DLL View : protected Field {
     /*!
      * child()と同じ
      * \since ver1.11
+     * \deprecated ver2.8〜
      */
-    View operator[](int index) const { return child(index); }
+    [[deprecated]]
+    View operator[](int index) const {
+        return child(std::to_string(index));
+    }
     /*!
      * \brief nameの最後のピリオドの前までを新しい名前とするField
      * \since ver1.11