+++ client/src/func.cc
@@ -114,6 +114,15 @@ int Func::index() const {
         return 0;
     }
 }
+const Func &Func::setIndex(int index) const {
+    auto func_info = dataLock()->func_store.getRecv(*this);
+    if (!func_info) {
+        throw std::invalid_argument("setIndex failed: Func not set");
+    } else {
+        (*func_info)->index = index;
+        return *this;
+    }
+}
 
 const Func &Func::setArgs(const std::vector<Arg> &args) const {
     auto func_info = setCheck()->func_store.getRecv(*this);