+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * installer.sh (#480, #481)
 * ValueElementRef, Value::at(), size() (#486)
 	* `Value::operator[](int)` の返り値型が変更になる
+* FuncInfoのindex (#487)
+	* Func::index(), setIndex()
 
 ## [2.7.0] - 2025-01-17
 ### Changed